+++ src/__tests__/math-spec.js
@@ -57,13 +57,12 @@ test("increment", t =>
  }`).then(outputIs(t, 3)));
 
 // Unary negation is not supported. Workaround: "return 0 - x;"
-test.skip("unary negation", t =>
+test("unary negation", t =>
   compileAndRun(`
   export function test(): i32 {
     let x: i32 = 2;
     return -x;
-  }`).then(outputIs(t, -2))
-);
+  }`).then(outputIs(t, -2)));
 
 test("uses precedence correctly", t =>
   compileAndRun(`
@@ -75,12 +74,20 @@ test("brackets", t =>
   compileAndRun(`
   export function test(): i32 {
     return 2 + (2 - 1);
-  }`)
-    .then(outputIs(t, 3))
-    .then(
-      // Slightly more complex brackets
-      compileAndRun(`
+  }`).then(outputIs(t, 3)));
+
+test("complex brackets", t =>
+  compileAndRun(`
+     export function test(): i32 {
+       return (2 * (3 - 1) - 1) / 3;
+     }`).then(outputIs(t, 1)));
+
+test("array index & math", t =>
+  compileAndRun(`
+    const memory: Memory = { 'initial': 1 };
     export function test(): i32 {
-      return (2 * (3 - 1) - 1) / 3;
-    }`).then(outputIs(t, 1))
-    ));
+      const x: i32[] = 0;
+      x[0] = 7;
+      x[0] = -x[0];
+      return x[0];
+    }`).then(outputIs(t, -7)));