+++ src/tokenizer/type/index.js
@@ -1,7 +1,7 @@
-const Trie = require("../../utils/trie");
-const token = require("../token");
-const Syntax = require("../../Syntax");
+import Trie from "../../utils/trie";
+import token from "../token";
+import Syntax from "../../Syntax";
 
 const supported = ["i32", "i64", "f32", "f64", "Function", "Memory", "void"];
 const trie = new Trie(supported);
-module.exports = token(trie.fsearch, Syntax.Type, supported);
+export default token(trie.fsearch, Syntax.Type, supported);